Fix warnings.

+++ source/base/net/firewall_manager.cc
@@ -35,7 +35,7 @@ FirewallManager::FirewallManager(const std::filesystem::path& application_path)
                                   IID_PPV_ARGS(&firewall_policy_));
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "CreateInstance failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "CreateInstance failed: " << SystemError::toString(static_cast<DWORD>(hr));
         firewall_policy_ = nullptr;
         return;
     }
@@ -43,7 +43,7 @@ FirewallManager::FirewallManager(const std::filesystem::path& application_path)
     hr = firewall_policy_->get_Rules(firewall_rules_.GetAddressOf());
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "get_Rules failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "get_Rules failed: " << SystemError::toString(static_cast<DWORD>(hr));
         firewall_rules_ = nullptr;
     }
 }
@@ -108,7 +108,8 @@ bool FirewallManager::addTcpRule(std::wstring_view rule_name,
     HRESULT hr = CoCreateInstance(CLSID_NetFwRule, nullptr, CLSCTX_ALL, IID_PPV_ARGS(&rule));
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "CoCreateInstance failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "CoCreateInstance failed: "
+                        << SystemError::toString(static_cast<DWORD>(hr));
         return false;
     }
 
@@ -125,7 +126,7 @@ bool FirewallManager::addTcpRule(std::wstring_view rule_name,
     firewall_rules_->Add(rule.Get());
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "Add failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "Add failed: " << SystemError::toString(static_cast<DWORD>(hr));
         return false;
     }
 
@@ -144,7 +145,7 @@ void FirewallManager::deleteRuleByName(std::wstring_view rule_name)
         HRESULT hr = rule->get_Name(bstr_rule_name.GetAddress());
         if (FAILED(hr))
         {
-            LOG(LS_WARNING) << "get_Name failed: " << SystemError::toString(hr);
+            LOG(LS_WARNING) << "get_Name failed: " << SystemError::toString(static_cast<DWORD>(hr));
             continue;
         }
 
@@ -172,7 +173,7 @@ void FirewallManager::allRules(std::vector<Microsoft::WRL::ComPtr<INetFwRule>>* 
     HRESULT hr = firewall_rules_->get__NewEnum(rules_enum_unknown.GetAddressOf());
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "get__NewEnum failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "get__NewEnum failed: " << SystemError::toString(static_cast<DWORD>(hr));
         return;
     }
 
@@ -181,7 +182,8 @@ void FirewallManager::allRules(std::vector<Microsoft::WRL::ComPtr<INetFwRule>>* 
     hr = rules_enum_unknown.CopyTo(rules_enum.GetAddressOf());
     if (FAILED(hr))
     {
-        LOG(LS_WARNING) << "QueryInterface failed: " << SystemError::toString(hr);
+        LOG(LS_WARNING) << "QueryInterface failed: "
+                        << SystemError::toString(static_cast<DWORD>(hr));
         return;
     }
 
@@ -191,7 +193,7 @@ void FirewallManager::allRules(std::vector<Microsoft::WRL::ComPtr<INetFwRule>>* 
         hr = rules_enum->Next(1, rule_var.GetAddress(), nullptr);
         if (FAILED(hr))
         {
-            LOG(LS_WARNING) << "Next failed: " << SystemError::toString(hr);
+            LOG(LS_WARNING) << "Next failed: " << SystemError::toString(static_cast<DWORD>(hr));
         }
 
         if (hr != S_OK)
@@ -207,7 +209,8 @@ void FirewallManager::allRules(std::vector<Microsoft::WRL::ComPtr<INetFwRule>>* 
         hr = V_DISPATCH(&rule_var)->QueryInterface(IID_PPV_ARGS(rule.GetAddressOf()));
         if (FAILED(hr))
         {
-            LOG(LS_WARNING) << "QueryInterface failed: " << SystemError::toString(hr);
+            LOG(LS_WARNING) << "QueryInterface failed: "
+                            << SystemError::toString(static_cast<DWORD>(hr));
             continue;
         }
 
@@ -215,7 +218,8 @@ void FirewallManager::allRules(std::vector<Microsoft::WRL::ComPtr<INetFwRule>>* 
         hr = rule->get_ApplicationName(bstr_path.GetAddress());
         if (FAILED(hr))
         {
-            LOG(LS_WARNING) << "get_ApplicationName failed: " << SystemError::toString(hr);
+            LOG(LS_WARNING) << "get_ApplicationName failed: "
+                            << SystemError::toString(static_cast<DWORD>(hr));
             continue;
         }
